requirements_sac.txt
# SAC Training Requirements
# Core RL libraries
stable-baselines3>=2.0.0
gymnasium>=0.29.0
torch>=2.0.0

# Communication
pyzmq>=25.0.0

# Utilities
numpy>=1.24.0
tensorboard>=2.13.0

# ROS2 dependencies (should already be installed)
# rclpy
# geometry_msgs
# sensor_msgs
# std_msgs
# tf2_ros